+++ src/main/java/lv/javaguru/student_anton_papin/lesson_1_homework/Homework_task_11.java
@@ -1,4 +1,4 @@
-package main.java.lv.javaguru.student_anton_papin;
+package main.java.lv.javaguru.student_anton_papin.lesson_1_homework;
 
 public class Homework_task_11 {
     public static void main(String[] args) {
