+++ src/main/java/lv/javaguru/student_anton_papin/lesson_2_homework/level_1/Lesson_2_homework_level_1_task_1.java
@@ -1,16 +1,17 @@
-package main.java.lv.javaguru.student_anton_papin;
+package main.java.lv.javaguru.student_anton_papin.lesson_2_homework.level_1;
 
 import java.util.Scanner;
 
-public class Lesson_2_homework_level_1_task_2 {
+public class Lesson_2_homework_level_1_task_1 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
-        System.out.println("Enter two numbers");
-        double numberA = scanner.nextDouble();
-        double numberB = scanner.nextDouble();
+        System.out.println("Enter two integers: ");
+        int numberA = scanner.nextInt();
+        int numberB = scanner.nextInt();
         System.out.println( numberA + numberB );
         System.out.println( numberA - numberB );
         System.out.println( numberA * numberB );
         System.out.println( numberA / numberB );
+
     }
 }
