+++ src/main/java/lv/javaguru/student_anton_papin/lesson_2_homework/level_2/Lesson_2_homework_level_2_task_7
