+++ src/main/java/lv/javaguru/student_anton_papin/lesson_2_homework/level_2/Lesson_2_homework_level_2_task_8.java
@@ -1,4 +1,4 @@
-package main.java.lv.javaguru.student_anton_papin;
+package main.java.lv.javaguru.student_anton_papin.lesson_2_homework.level_2;
 
 import java.util.Scanner;
