+++ src/main/java/lv/javaguru/student_anton_papin/lesson_2_homework/level_3/Lesson_2_homework_level_3_task_10.java
@@ -1,4 +1,4 @@
-package main.java.lv.javaguru.student_anton_papin;
+package main.java.lv.javaguru.student_anton_papin.lesson_2_homework.level_3;
 
 import java.util.Scanner;
