+++ src/main/java/lv/javaguru/student_anton_papin/lesson_2_homework/level_5/Lesson_2_homework_level_5_task_15.java
@@ -1,4 +1,4 @@
-package main.java.lv.javaguru.student_anton_papin;
+package main.java.lv.javaguru.student_anton_papin.lesson_2_homework.level_5;
 
 public class Lesson_2_homework_level_5_task_15 {
     public static void main(String[] args) {
