+++ src/main/java/lv/javaguru/student_anton_papin/lesson_2_homework/level_6/Lesson_2_homework_level_6_task_23.java
@@ -1,4 +1,4 @@
-package main.java.lv.javaguru.student_anton_papin;
+package main.java.lv.javaguru.student_anton_papin.lesson_2_homework.level_6;
 
 public class Lesson_2_homework_level_6_task_23 {
