+++ src/main/java/lv/javaguru/student_anton_papin/lesson_2_homework/level_6/Lesson_2_homework_level_6_task_24.java
@@ -1,4 +1,4 @@
-package main.java.lv.javaguru.student_anton_papin;
+package main.java.lv.javaguru.student_anton_papin.lesson_2_homework.level_6;
 
 public class Lesson_2_homework_level_6_task_24 {public static void main(String[] args) {
     System.out.println("I wanna coffee");
